Initial Refactoring and migration to fastapi, which enables a simpler way to do ubiquiti
