+++ Containerfile
@@ -1,4 +1,4 @@
-FROM docker.io/library/python:3.9.9-slim-bullseye
+FROM docker.io/library/python:3.12.7-slim-bullseye
 
 RUN python3 -m venv /opt/venv
 
@@ -7,13 +7,16 @@ COPY etc/webserver_requirements.txt requirements.txt
 RUN . /opt/venv/bin/activate && pip install -r requirements.txt
 
 WORKDIR /opt/webserver/cgi-bin/
-COPY cgi-bin/dns_server.py /opt/webserver/cgi-bin/
+COPY cgi-bin/controller.py /opt/webserver/cgi-bin/
 COPY cgi-bin/gunicorn_config.py /opt/webserver/cgi-bin/
 COPY cgi-bin/wsgi.py /opt/webserver/cgi-bin/
 
 WORKDIR /opt/webserver/etc/
 COPY etc/config.ini /opt/webserver/etc/
 
-#PYTHONPATH=../../dns_admin/lib/ 
+WORKDIR /opt/webserver/cgi-bin/
+COPY lib/. /opt/webserver/lib/
+
+#PYTHONPATH=../../dns_admin/lib/
 WORKDIR /opt/webserver/
-CMD . /opt/venv/bin/activate && exec gunicorn -c cgi-bin/gunicorn_config.py --reload  --chdir cgi-bin wsgi:app
+CMD . /opt/venv/bin/activate && exec gunicorn -k uvicorn.workers.UvicornWorker -c cgi-bin/gunicorn_config.py --reload  --chdir cgi-bin wsgi:app
