+++ etc/webserver_requirements.txt
@@ -1,19 +1,21 @@
-aniso8601==9.0.1
-certifi==2021.10.8
-charset-normalizer==2.0.7
-click==8.0.3
-Flask==2.0.2
-Flask-RESTful==0.3.9
-gunicorn==20.1.0 
-idna==3.3
-importlib-metadata==4.8.2
-itsdangerous==2.0.1
-Jinja2==3.0.3
-MarkupSafe==2.0.1
-pytz==2021.3
-requests==2.26.0
-six==1.16.0
-typing-extensions==3.10.0.2
-urllib3==1.26.7
-Werkzeug==2.0.2
-zipp==3.6.0
+annotated-types==0.7.0
+anyio==4.6.0
+certifi==2024.8.30
+charset-normalizer==3.3.2
+click==8.1.7
+fastapi[standard]==0.115.0
+gunicorn==23.0.0
+h11==0.14.0
+idna==3.10
+Jinja2==3.1.4
+MarkupSafe==2.1.5
+packaging==24.1
+pydantic==2.9.2
+pydantic_core==2.23.4
+requests==2.32.3
+sniffio==1.3.1
+starlette==0.38.6
+typing_extensions==4.12.2
+urllib3==2.2.3
+uvicorn==0.31.0
+pyclean
