+++ pyrightconfig.json
@@ -1,0 +1,24 @@
+{
+  "typeCheckingMode": "basic",
+  "reportReturnType": "warning",
+  "reportArgumentType": "warning",
+  "reportOperatorIssue": "warning",
+  "reportImportCycles": "warning",
+  "reportDuplicateImport": "warning",
+  "venvPath": ".",
+  "venv": "venv",
+  "pythonPath": "venv/bin/python",
+  "extraPaths": ["./lib"],
+  "verboseOutput": true,
+  "executionEnvironments": [
+    {
+      "root": "./dns_admin",
+      "venvPath": "./dns_admin/",
+      "venv": "venv",
+      "extraPaths": ["./dns_admin/lib"]
+    },
+    {
+      "root": "./"
+    }
+  ]
+}
